Add main page link navbar
