+++ app/Http/Controllers/PublicController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Actions\GetAliasSubdomainForPublic;
+use App\Models\MainPage;
 use App\Models\Padalinys;
 use Illuminate\Support\Facades\Cache;
 use Inertia\Inertia;
@@ -23,7 +24,9 @@ class PublicController extends Controller
         $this->padalinys = Padalinys::where('alias', $alias)->first();
 
         // Subdomain and alias won't be different, except when alias = 'vusa', then subdomain = 'www'
-        Inertia::share('padalinys', $this->padalinys->only(['id', 'shortname', 'alias', 'type']) + ['subdomain' => $subdomain]);
+        Inertia::share('padalinys', $this->padalinys->only(['id', 'shortname', 'alias', 'type']) +
+            ['subdomain' => $subdomain]
+        );
     }
 
     protected function getBanners()
@@ -40,5 +43,12 @@ class PublicController extends Controller
         });
 
         Inertia::share('banners', $banners);
+    }
+
+    protected function getPadalinysLinks() {
+
+        $mainPage = MainPage::where([['padalinys_id', $this->padalinys->id], ['lang', app()->getLocale()]])->get();
+
+        Inertia::share('padalinys.links', $mainPage);
     }
 }
