+++ resources/css/app.css
@@ -135,3 +135,20 @@
 html.dark {
   color-scheme: dark;
 }
+
+/* .wrapper {
+  display: grid;
+  grid-template-columns:
+    1fr
+    min(65ch, 100%)
+    1fr;
+}
+
+.wrapper > * {
+  grid-column: 2;
+}
+
+.full-bleed {
+  width: 100%;
+  grid-column: 1 / 4;
+} */
